Initial add.
This sort of limps along, playing the attract mode and the Bower scene
more or less correctly, but I need to spend some time fine-tuning the
timings in any case.

It obviously needs more levels added, a level manager, timing fixes,
audio, game controller support, and a libretro platform layer before I'm
done...and that's just for Dragon's Lair! I'd love to get other QTE-style
games in here too, but also add a way to draw over the video for games like
Cobra Command, etc.

But still, I'm pretty happy with the progress on this already.