Bunch More Work.
This now allows for much more robust scene definitions, and has the
flaming ropes room filled in as an example, since it has a branching
piece of game logic (you either go right to the ropes or you have the
sliding platform before you go to the ropes). The timings aren't perfect
in this yet, but we're getting there.

Managing video during seeking is much improved too, although the act of
seeking in the Theora video itself still needs some work.