Tweaked video frames to millisecond calculation.
This seems to be more correct; the original ROM treats the video as
running at 23.976fps, and then I had to subtract out a little more
than six seconds to make it line up correctly, which I assume accounts
for the ROM check messages, etc, at the start of the disc.

This seems to make everything from the flaming ropes (2:22 into the
disc) to the literal dragon's lair (20:00 into the disc) show up
at the expected frame position, so it certainly _seems_ correct.