audio: Greatly improve audio output after a seek.
We will leave it to the platform layer to decide if the mainloop is
not keeping up with audio playback and dump its queue, but currently,
the SDL backend doesn't (RetroArch presumably handles this, though...?).

But in practice, even if the audio goes out of sync, we're going to
flush the audio entirely and start a new scene every couple of
seconds in most laserdisc games anyhow.