lair: Sprinkled a few `kills_player=true` around.
God mode chooses the first choice that leads to a sequence that
isn't marked `kills_player=true`, but the reversed long falling
platform and the fire room both have choices that lead to a
non-interactive sequence that, on timeout, leads to a
player-killing sequence.

Mark those sequences as fatal, too, to fix god mode. In normal
play, the game doesn't look at `kills_player` unless there's no
`nextsequence` available, so it won't end the scene early to
be there.